browser: add 30s delay before start
Neither Weston nor Grafana are ready so fast. This delay prevents the
browser from crashing due to Weston not being ready, and the page from
being rendered as 404 due to Grafana not being ready.

There could be more elegant ways to wait that are not based on an
arbitrary delay, but the delay approach is super easy and the startup
time can be slower than optimal.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>