compose: explicitly define platforms
The project uses containers from arm and arm64. When those are built
locally and pushed to Docker Hub as single-platform images, they run ok
even on a mismatch platform, only printing a warning.

On the other hand, when built as multi-platform images on the GitHub
Actions workflows, they fail to run on mismatch platforms because the
corresponding ones are not found on the manifest.

By explicitly defining the platforms, all containers are able to run
even when pushed to Docker Hub as multi-platform images.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>