data: don't ignore data point when unavailable
Ignoring the entire data point when a cURL fails (in other words, when
unavailable) leads to a panel with "no data" which is misleading.

There actually is data, and maybe it just happens it is "unavailable" to
all builds. Or it is unavailable to a single build.

To fix this, save the data points even when cURL request fails, and show
data as unavailable in the charts.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>