browser: add carousel mode to loop through URLs
Add a mode that makes it possible to switch URL.

Among several possible methods to use, this one has the switch logic
embedded into the browser container. It relies on Docker's
"restart=always" capability to bring the container up again after
purposefully killing the PID 1.

A random website is chosen when the container starts.

Closes #8

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>