browser: get-next-url: add concept of app selection
Before this commit, the script selects a random hardcoded URL. This is
not very flexible if an app may provide a random URL value among
different possible ones without messing with the "app randomness".

The implementation on this commit is backward compatible with the
previous ones. If files for the tdx_br_photo app are not available, the
Toradex logo is returned.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>