influxdb: limit container RAM allocation
When the DB starts to get too much data, InfluxDB gets really
memory-hungry. This leads to unexpected reboots as the OOM seems to be
killing the Docker daemon, which is watched by SystemD watchdog.

This commit does not fix the underlying issue with InfluxDB memory
consumption, but it prevents the unexpected reboots.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>