vscode: make update task more reliable
The update task has has some false negative updates in the past. Make it
a bit harder for that to happen by improving the regex that matches the
places that should be updated.

Furthermore, print the diff for the user instead of just asking, so that
a better UX is provided.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>